# server/docs/swagger.yaml
basePath: /
definitions:
...
  v1.CommentRequest:
    properties:
      comment:
        description: actual content of comment, review
        type: string
      comment_type:
        description: review, etc.
        type: string
      episode_number:
        type: integer
      score:
        description: only required for reviews
        type: integer
      season_number:
        description: only for tvshows, when commenting on a particular episode
        type: integer
      title:
        type: string
    required:
    - comment_type
    type: object
...
  v1.V1ErrorResponse:
    properties:
      error:
        example: error message
        type: string
      status:
        enum:
        - success
        - error
        example: error
        type: string
    type: object
  v1.V1SuccessResponse:
    properties:
      data: {}
      status:
        enum:
        - success
        - error
        example: success
        type: string
    type: object
...
  view.CommentObject:
    properties:
      comment:
        description: actual content of comment, review
        type: string
      comment_id:
        type: integer
      comment_type:
        type: string
      created_at:
        type: string
      is_public:
        type: boolean
      record_id:
        type: integer
      score:
        type: integer
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:2323
info:
  contact: {}
  description: |-
    V1 Hound Media Server API Documentation
    While in beta, please expect breaking api changes in short/no notice
  title: Hound API Documentation V1
  version: beta-0.1.0
paths:
...
  /api/v1/comments/{id}:
    delete:
      consumes:
      - application/json
      parameters:
      - description: Comment ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Delete a comment
      tags:
      - Comments
...
  /api/v1/movie/{id}/comments:
    get:
      consumes:
      - application/json
      parameters:
      - description: Media ID
        example: tmdb-1234
        in: path
        name: id
        required: true
        type: string
      - description: Comment Type
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/view.CommentObject'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Get comments for a movie
      tags:
      - Comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Media ID
        example: tmdb-1234
        in: path
        name: id
        required: true
        type: string
      - description: Comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/v1.CommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Post a comment for a movie
      tags:
      - Comments
...
  /api/v1/tv/{id}/comments:
    get:
      consumes:
      - application/json
      parameters:
      - description: Media ID
        example: tmdb-1234
        in: path
        name: id
        required: true
        type: string
      - description: Comment Type
        in: query
        name: type
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  items:
                    $ref: '#/definitions/view.CommentObject'
                  type: array
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Get comments for a TV show
      tags:
      - Comments
    post:
      consumes:
      - application/json
      parameters:
      - description: Media ID
        example: tmdb-1234
        in: path
        name: id
        required: true
        type: string
      - description: Comment
        in: body
        name: comment
        required: true
        schema:
          $ref: '#/definitions/v1.CommentRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            allOf:
            - $ref: '#/definitions/v1.V1SuccessResponse'
            - properties:
                data:
                  type: object
              type: object
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/v1.V1ErrorResponse'
      summary: Post a comment for a TV show
      tags:
      - Comments
...
swagger: "2.0"